requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
python-dotenv==1.0.1
gunicorn==22.0.0
reportlab==4.1.0
google-genai
Pillow
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
scikit-learn
numpy
cryptography